[#105] refactored code a little
