+++ src/ml_service/WordVectors/FrenchVectors.py
@@ -1,0 +1,73 @@
+# -*- coding: utf-8 -*-
+import numpy
+import os
+from gensim.models.keyedvectors import KeyedVectors
+from nltk.corpus import stopwords
+
+
+# #################################################
+# LOAD FROM .bin
+# -------------------------------------------------
+# return word_vector_model
+def load_from_bin():
+    try:
+        print("Loading word vector file... May take a few seconds")
+        __script_dir = os.path.abspath(__file__ + r"/../../")
+        __rel_path = r'WordVectors/non-lem.bin'
+        file = os.path.join(__script_dir, __rel_path)
+        model = KeyedVectors.load_word2vec_format(file, binary=True)
+        print("Loading complete")
+        return model
+    except BaseException:
+        print("Download French Vector model first")
+
+def get_stop_words():
+    return stopwords.words('french') + \
+        [',', ';', '.', '!', '?', 'c', '(', ')', 'ainsi',
+            'alors', 'au', 'aucuns', 'aussi', 'autre', 'avant', 'avec',
+            'avoir', 'bon', 'car', 'ce', 'cela', 'ces', 'ceux',
+            'chaque', 'ci', 'comme', 'comment', 'dans', 'des', 'du', 'dedans',
+            'dehors', 'depuis', 'devrait', 'doit', 'donc', 'dos', 'début',
+            'elle', 'elles', 'en', 'encore', 'essai', 'est', 'et', 'eu', 'fait',
+            'faites', 'fois', 'font', 'hors', 'ici', 'il', 'ils', 'la', 'le', 'les',
+            'leur', 'là', 'ma', 'maintenant', 'mais', 'mes', 'mine', 'moins',
+            'mon', 'mot', 'même', 'ni', 'nommés', 'notre', 'nous', 'ou',
+            'où', 'par', 'parce', 'pas', 'peut', 'peu', 'plupart', 'pour',
+            'pourquoi', 'quand', 'que', 'quel', 'quelle,''quelles', 'quels',
+            'qui', 'sa', 'sans', 'ses', 'seulement', 'si', 'sien,''son', 'sont',
+            'sous', 'soyez', 'sur', 'ta', 'tandis', 'tellement', 'tels',
+            'tes', 'ton', 'tous''tout', 'trop', 'très', 'tu', 'voient',
+            'vont', 'votre', 'vous', 'vu', 'ça', 'étaient', 'état', 'étions', 'été', 'être',
+         ]
+
+class FrenchVectors:
+    word_vectors = load_from_bin()
+    custom_stop_words = get_stop_words()
+    Word_Vector_Size = 500
+
+    def __init__(self):
+        pass
+
+    '''
+    ----------------------------------------------------------------
+    Vectorize Sentence
+    ----------------------------------------------------------------
+    
+    Adds vectors together and divides by number of words
+    
+    word_list <list[String]>: list of words representing a sentence
+    return: numpy array
+    '''
+    @staticmethod
+    def vectorize_sent(word_list):
+        vector = numpy.zeros(FrenchVectors.Word_Vector_Size)
+        num = 0
+        for word in word_list:
+            try:
+                if word in FrenchVectors.custom_stop_words:
+                    continue
+                vector = numpy.add(vector, FrenchVectors.word_vectors[word])
+                num += 1
+            except KeyError:
+                pass
+        return numpy.divide(vector, num)
